Use The Graph for claiming withdraws (#172)
In order to fully decommission mongoDB we need to be able to run the e2e tests without it. Right now we still rely on it in the claim withdraw script to compute the merkle inclusion proof when claiming a requested withdraw.

This PR switches that script from using Mongo to using the graph. I decided to use the Rest API (similar to how we already do it in the e2e tests since this is the easiest to parse)

### Test Plan

Patch dex-services repo with this commit, see withdraw e2e test pass.