[FlakyTest] Query amount of seconds remaining in current batch (#194)
We are seeing a bunch of flaky tests. In the errors, I can always see the following exception:

```
Error: Returned error: VM Exception while processing transaction: revert -- Reason given: Solutions are no longer accepted for this batch.
```

e.g. https://travis-ci.org/gnosis/dex-contracts/builds/583550493

I believe the reason for them is a timing issue in the way we forward tome for the auction to be over. We always forward the entire batch time (5 minutes). This means that if we are currently very close to the end of a  batch (e.g. less than 5s) and the time it takes to forwards time and then submit the solution is larger than that, we are already in the `k+2`nd batch.

My proposed fix is to wait for the auction to close by checking how much time is left in the current batch. To do this, I'm exposing a new view method, which I believe will be useful also for general interactions (e.g. when we place an order in the UI, we might want to know how long until we can expect people trading against it).

### Test Plan

No more flaky unit tests.