Return empty bytes instead of reverting when auction is empty (#214)
Right now we are seeing a lot of errors in the driver. This is because the contract throws when `getEncodedAuctionElements` is called on an empty auction (after deploying it).

We should probably just return an empty byte array in this case.

### Test Plan

pull this in dex-services and see that we don't get the weird error about the empty set any more.