Prepare first NPM package (#323)
Hi! Nick and I were reviewing how to share JS pieces among the projects. And the current proposal we are going to work on is:

    dex-contract: ABI, TS definitions, and even utils close to the contract, like unpack orders
    dex-js: utilities that are used in Front ends, bots, services, adds some goodies like formatting, etc. Also re-export the utils from dex-contract so dex-js is the only thing you need to integrate using JS

This PR gets dex-contract package ready for publishing to NPM by fixing up the package.json and moving all dependencies into devDependencies.

### Test Plan

CI is undisturbed

### Commit History

* Prepare first NPM package

* 0.0.1

* Delete unnecesary yarn.lock file

* updated package-lock with for new dev dependencies

* alphabetically order dependencies

* add .history back in

* remove dependencies from NPM package completely