[BugFix] using `user` in withdraw (#399)
In the function withdraw, we used msg.sender and the variable user.

Bug found by Adam.

I changed the tests, so that they would spot it.

How could this happen? Analysis:
The relevant change was made in this PR:
#233

First, I implemented the fix only with msg.sender and without any use of the "user", it got approved. And then I did the merge with master. Unfortunately, I did not pay enough attention to this merge and... the "user" was introduced in withdraw, but not in the new lines of withdraw().

What can we learn about?

being more careful with merges - definitively
approving after merges? - this is not practical, I guess.