Emit event on SolutionSubmission (#469)
Worst case gas cost: 20k

This will allow us to track information such as prices, different ingredients of the total objective value and fees burnt.

### Test Plan

Benchmarked impact on the worst case ring trade example. 

It turns out we need to provide >6.8M gas although we in the end only end up using 5.2M. This is because the gas refund for releasing storage is only granted at the end of the transaction and the provided gas needs to be large enough to be non-negative at all times throughout the tx.