[BatchExchangeViewer] Wrapper contract to show current order book (#535)
Part of #528 

This PR introduces a BatchExchange wrapper contract only consisting of view functions. The goal is for this contract to allow slicing and dicing of the on chain data while being a completely separate and thus easy to change and update as we see fit.

It can also be serve as a code sharing facility between our multi-language repos (python, rust, js, etc).

As a first step I introduce two functions in paginated and unpaginated form that both fetch the current order book in all dimensions. One fetches the one that is currently still open for order submission, the other one fetches the orderbook that is currently being solved.

In the next step I will introduce a filtering option to get the orderbook only for certain tokens.

cc @anxolin @alfetopito @W3stside @Velenir as this might be useful on the frontend as well.

### Test Plan

Updated unit tests. To use the wrapper on rinkeby and fetch the entire order book in a single query do the following:

1. `export PK=< your rinkeby eth funded private key >`
2. `yarn prepack`
3. `npx truffle migrate -f 4 --network rinkeby`
4. `npx truffle console --network rinkeby`

```js
const zero_address = "0x0000000000000000000000000000000000000000"
const viewer = await BatchExchangeViewer.deployed()
result = await viewer.contract.methods.getFinalizedOrderBookPaginated(zero_address, 0, 250).call({gas: 100000000})
result.elements.length / 112 / 2 // should be less than 250
```