[Viewer] Allow sellTokenBalance to be computed for a certain batchId (#569)
Reported by @koeppelmann 

When we query the current order book we are getting the sellTokenBalances that are valid for the auction that is currently being sold. However, when visualizing the orderbook for the purpose of predicting at what price we should place an order to be matched in the next auction, we need to consider the balances that will be valid in the next auction.

For this we need to potentially add pendingDeposits and subtract pendingWithdraws.

This PR allows to specify a sellTokenBalanceTargetBatchId into the main filter function and will adjust the returned sellTokenBalance for this targetBatch. This means `getFinalizedOrderBook` will use the current batch ID whereas `getOpenOrderBook` will use the next batch ID.

### Test Plan

Added unit test explaining the behavior.