Increase memory for node to fix master branch (#591)
Travis master is failing because it is currently running out of memory (cf. https://travis-ci.org/github/gnosis/dex-contracts/builds/662931725?utm_medium=notification&utm_source=github_status)

This PR increases the available RAM for our node process to 4GB which will unblock us. I will take a look at where we are using so much memory.

After increasing memory, it was still failing for testing the gas cap on our worst case scenario (reversion and reaplication of a solution with 30 trades). This looks to be related to the ganache version (it works on Ganache CLI v6.7.0 but fails on Ganache CLI v6.9.1 locally).

Travis is using @latest und thus may have started failing without us noticing via and explicit PR. I'm therefore making the dependency explicit inside our package.json (so that we will create a PR for updating it, which makes it easier to pinpoint failures due to version upgrades) and increase the gas limit to 6M in order to make the test pass.

### Test Plan

travis