Introducing an orderbook class (typescript) (#581)
The goal of this PR is to lay the foundation of allowing typesafe logic to be added to the dex-contracts repo. The first use case will be an Orderbook class, which can be created from bids and asks and for now simply serialized into a json format that can be easily visualized (it's the same format I'm using here: https://6jxh4.csb.app/)

Going forward this class will receive logic such as inverting an orderbook (bids become asks and vice versa), "adding" orderbooks (combining two books on the same pair into one) and "multiplying" (turning A:B & B:C into A:C). All this will be useful for visualization of order data as well as for estimating what limit price an order should be using in order to have a good chance of being matched.

The typescript compilation and tests are run as part of travis and we will be able to depend on the compiled js resources in our js scripts.

### Test Plan

`yarn run test-ts` and making sure that after running yarn build we can depend on `/typescript/orderbook.js` in the good old javascript truffle scripts.