[Orderbook] logic to invert orderbook (#592)
This PR introduces logic, which allows to "invert" a given orderbook. This will e.g. turn an ETH/DAI book into a DAI/ETH book. 

This can be useful as instead of representing each token pair twice (A/B & B/A) we can instead chose a canonical representation (e.g. by alphabet, thus A/B) and still use the inverse orderbook (e.g. if I want to compute the book between B/C and would like to use B/A * A/C to get there)

### Test Plan

added a unit test