[Fraction] Allow to convert large fractions to number (#605)
I was running into issue when trying to print prices after some operations saying that number can only safely store up to 53 bits.

This is when we multiply a bunch of prices that have no common denominator. In this case it probably makes sense to shift the numbers by the amount of bits they have too many and then divide the remaingin (approximated amount). Since this is only used to get a number approximation that should be fine.

### Test Plan

Added unit test with two large primes.