[Orderbook] incorporating fee (#610)
This PR starts incorporating the trading fee that we collect on our exchange into the price estimation algorithm.

Whenever adding a bid/offer we deduct the fee from both volume and the price. This should ensure that when multiplying orders with one another, the resulting order is accounting for the two hops and thus 2*fee that it would incur.

### Test Plan
Added unit tests. Made sure we are still getting reasonable liquidity when running the e2e script