[Viewer] Stop fetching more sub-pages when running low on gas (#631)
We are running into issues when paginating over filtering BatchExchangeViewer methods such as `getOpenOrderBookPaginated`. The reason is that in order to fill an entire pageSize (e.g. 100) we might have to do many subqueries to the unfiltered `getEncodedOrdersPaginated`. This can make us run out of gas even for small page sizes if one account (e.g. our e2e test account on rinkeby or the EFrontier account on mainnet) has a lot of expired orders which don't match our filter.

This PR therefore changes the logic to make `pageSize` a `maxPageSize` and measure in each loop how much gas was used for the subquery. If the last call used more than half of the remaining gas it is unlikely that another iteration is not possible and we return early with what we have found so far.

Note, that this requires to add another flag to the returned page information as we can no longer count the elements returned and compare it to the page size.

An alternative implementation could hard-code the page size for the subquery based on the amount of `gasleft()` but this would require careful estimation of the gas cost which and might be brittle to changes in gas cost per operation (cf. Istanbul increase for SLOAD operation).

### Test Plan

To see the difference between the old and the new implementation, do the following:

```bash
npx truffle migrate -f 3 --network rinkeby
npx truffle console --network rinkeby
```
In the console:
```js
let viewer = await BatchExchangeViewer.deployed()
let old = await BatchExchangeViewer.at("0x3Cc80340Cb9C2C865c44b0f491968ACb9EaD2b48") // current version on rinkeby
await old.getOpenOrderBookPaginated([], "0x0000000000000000000000000000000000000000", 0, 50) // this will run out of gas
await viewer.getOpenOrderBookPaginated([], "0x0000000000000000000000000000000000000000", 0, 50) // this will work
```