[TS] Refactor test resources to typescript (#629)
This PR moves the large test scenarios that we implemented from js to typescripts (since those are used in a bunch of classes that I would like to convert next).

This was a little bit more involved. While the code was very well documented and even had voluntary type annotations, it's also pretty heavy in logic. Also some of the annotations were out of date, which emphasises the benefit of having them checked at compile time.

I'm not sure what I should do with the type annotations in the comments. Technically they should no longer be needed.

### Test Plan

CI