[Fraction] Allow toNumber on parts that overflow floats (#646)
Currently when looking at the sUSD/USDC order pair on https://czii9.csb.app/ we can see more bid liquidity when hops=1 (>35k)) than for hops=2 (<25k)

This is counterintuitive as the available volume should strictly grow the more hops we use.

I debugged this yesterday and found that the volume for hops = 2 is `null` for a bunch of offers. This stems from the fact that we are multiplying a bunch of large fractions to the point where numerator/denominator no longer fit into floats. Thus our `toNumber()` implementation returns `NaN` which gets translated to `null`.

This PR fixes this by shifting numerator and denominator right until both of them fit into floats. 

Some local testing suggested that this should fix the anomaly that can be seen on the graph above.

### Test Plan

Unit tests + yalcing this change in the price estimator and see that the bids returned for hops = 1 are a subset of the bids returned for hops = 2