[BatchExchangeViewer] Allow token filter for getEncodedOrders (#663)
This PR addresses an issue that @e00E mentioned in slack. When a user has an invalid sellBalance (e.g. from listing a fake ERC20 token that has no reasonable supply) our current way of fetching the on-chain orderbook via a smart contract would fail.

This PR adds a unit-test outlining the issue and implements a filtered version of `getEncodedOrdersPaginated`. For this we reimplement the encode logic as well as the fetching logic inside BatchExchangeViewer but wrap it around a token filter check before invoking getBalance.

Note, that there is no good way of computing the number of orders a certain user has, thus we need to "try" fetching the next order of a given use and inspect the return flag if the static call succeeded.

### Test Plan

Added a unit test and did some high tenderly testing that the implementation is using around the same gas as the current one.