[BatchExchangeViewer] Serialize orderId when querying orders with filter (#668)
It would be nice to use the filtered orderbook for querying inside dex-services. This should reduce the amount of data we are sending over the wire and thus speed up the fetching time.

The thing that is preventing this at the moment is that from a filtered orderbook we cannot reconstruct the orderId (it's not incrementing as there might be gaps) and thus we cannot use this order information for `submitSolution`.

This PR changes that. It appends the orderId (index of this order in all of the user's orders) to the bytes returned from any of the filtering methods (such as `getOpenOrderBook`, `getFinalizedOrderBook`). It also updated the decode methods to comply with the new format.

### Test Plan

Adjusted unit test to assert filter logic based on orderId.