[Events Orderbook] Refactor `applyEvent` Logic (#725)
The refactor that was promised! This PR refactors `applyEvents` switch statement to be more "dumb" by forwarding each event to their individual functions.

This has the added benefit of using the `Event` types so that we have type safety on the event `returnValues` property.

### Test Plan

This PR is just a refactor so unit tests should verify that nothing broke.