[Easy] Remove strict withdraw request check (#761)
This PR removes a strict withdraw request check that was making sure a previous withdraw request would not get overwritten. While this cannot happen for matured withdraw requests, it is possible for future withdraw requests that have not yet matured to be overwritten. Since the state currently does note get batch index information with events, we just remove the check for now and leave a TODO to add it back in later once we provide batch information with the events.

### Test Plan

```
ETHEREUM_NODE_URL=https://some.rinkeby.node yarn ts-streamed-orderbook
```

Note that the test was slightly modified. It looks like something might have changed in `BN` as it was producing numbers with extra leading 0's for some reason and causing the tests to fail on benign errors:
```
             "words": [
               46182994
               463411
               13477242
      +        0
      +        0
      +        0
      +        0
      +        0
      +        0
      +        0
      +        0
             ]
           }
           "user": "0x740a98f8f4fae0986fb3264fe4aacf94ac1ee96f"
           "validFrom": 5271160
```