[Viewer] Convenience method to get tokenInfo for a tokenID (#860)
I'm planning to use this method in dex-services when we want to get the symbol and decimals for each token. At the moment it would require three calls:

1. turn tokenId into tokenAddress
2. Get symbol for tokenAddress
3. Get decimals for tokenAddress

While 2./3. could happen in one batch they depend on the result of query 1. With this change we can query token Info atomically in one call.

I thought about allowing to pass in an array and return an array of symbols, decimals and addresses, however the return type string[] is only available in the experimental ABI. The alternative would be to byte encode this information similar to how we do for order information, however this did not feel worth the overhead it would add consuming this method (and multiple tokenId requests can be batched into one roundtrip).

### Test Plan

Added unit test.